dwc2: use the nonblock argument in submit_int_msg
An USB 1.1 keybaord connected to dwc2 does not report status until it
changes. With this patch you can enable keyboard by pressing a key while
USB devices are probed. Without a keypress no state is reported an the
probe times out. We don't want to wait for a keypress or timeout while
polling for keypresses so implement the nonblock variant that exits early.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>